feat!
Se actualiza tipo de dato del cuitcuit a long para poder trabajar con los 11 digitos del mismo

+++ ESCMB.HybridDDDArchitecture/ESCMB.Application/UseCases/Client/Commands/CreateClient/CreateClientCommand.cs
@@ -8,7 +8,7 @@ namespace ESCMB.Application.UseCases.Client.Commands.CreateClient
         [Required]
         public string Apellido { get; set; }
         [Required]
-        public int CuitCuil { get; set; }
+        public long CuitCuil { get; set; }
         [Required]
         public string Email { get; set; }
         [Required]
